cobranza por dias fechas actualizadas en varias partes migraciones actualizadas

+++ database/migrations/2025_04_30_192414_create_payments_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->date('payment_date');
             $table->date('payment_due');
             $table->unsignedBigInteger('collector')->nullable(); // Puede ser nulo
-            $table->bigInteger('amount');
+            $table->float('amount');
             $table->boolean('pay_full')->default(false);
             $table->timestamps();
         });
